+++ backend/src/models/event_models.rs
@@ -5,7 +5,7 @@ use diesel::*;
 
 use crate::{schema::{events, event_instances}, db_connection::PgPooledConnection};
 
-pub fn get_event(event_id: i32, conn: &mut PgPooledConnection,) -> Result<Event, Status> {
+pub fn get_event(event_id: i64, conn: &mut PgPooledConnection,) -> Result<Event, Status> {
     events::table
         .select(events::all_columns)
         .filter(events::id.eq(event_id))
@@ -13,7 +13,7 @@ pub fn get_event(event_id: i32, conn: &mut PgPooledConnection,) -> Result<Event,
         .map_err(|_| Status::new(Code::NotFound, "event_not_found"))
 }
 
-// pub fn get_group_event(group_id: i32, event_id: i32, conn: &mut PgPooledConnection,) -> Result<GroupEvent, Status> {
+// pub fn get_group_event(group_id: i64, event_id: i64, conn: &mut PgPooledConnection,) -> Result<GroupEvent, Status> {
 //     group_posts::table
 //         .select(group_posts::all_columns)
 //         .filter(group_posts::group_id.eq(group_id))
@@ -24,8 +24,8 @@ pub fn get_event(event_id: i32, conn: &mut PgPooledConnection,) -> Result<Event,
 
 #[derive(Debug, Queryable, Identifiable, AsChangeset)]
 pub struct Event {
-    pub id: i32,
-    pub post_id: i32,
+    pub id: i64,
+    pub post_id: i64,
     pub info: serde_json::Value,
     pub created_at: SystemTime,
     pub updated_at: Option<SystemTime>,
@@ -34,15 +34,15 @@ pub struct Event {
 #[derive(Debug, Insertable)]
 #[diesel(table_name = events)]
 pub struct NewEvent {
-    pub post_id: i32,
+    pub post_id: i64,
     pub info: serde_json::Value,
 }
 
 #[derive(Debug, Queryable, Identifiable, AsChangeset)]
 pub struct EventInstance {
-    pub id: i32,
-    pub event_id: i32,
-    pub post_id: Option<i32>,
+    pub id: i64,
+    pub event_id: i64,
+    pub post_id: Option<i64>,
     pub info: serde_json::Value,
     pub starts_at: SystemTime,
     pub ends_at: SystemTime,
@@ -53,8 +53,8 @@ pub struct EventInstance {
 #[derive(Debug, Insertable)]
 #[diesel(table_name = event_instances)]
 pub struct NewEventInstance {
-    pub event_id: i32,
-    pub post_id: Option<i32>,
+    pub event_id: i64,
+    pub post_id: Option<i64>,
     pub info: serde_json::Value,
     pub starts_at: SystemTime,
     pub ends_at: SystemTime,
@@ -62,10 +62,10 @@ pub struct NewEventInstance {
 
 // #[derive(Debug, Queryable, Identifiable, AsChangeset)]
 // pub struct GroupEvent {
-//     pub id: i32,
-//     pub group_id: i32,
-//     pub event_id: i32,
-//     pub user_id: i32,
+//     pub id: i64,
+//     pub group_id: i64,
+//     pub event_id: i64,
+//     pub user_id: i64,
 //     pub group_moderation: String,
 //     pub created_at: SystemTime,
 //     pub updated_at: Option<SystemTime>,
@@ -73,23 +73,23 @@ pub struct NewEventInstance {
 // #[derive(Debug, Insertable)]
 // #[diesel(table_name = group_posts)]
 // pub struct NewGroupEvent {
-//     pub group_id: i32,
-//     pub event_id: i32,
-//     pub user_id: i32,
+//     pub group_id: i64,
+//     pub event_id: i64,
+//     pub user_id: i64,
 //     pub group_moderation: String,
 // }
 
 // #[derive(Debug, Queryable, Identifiable, AsChangeset)]
 // pub struct UserEvent {
-//     pub id: i32,
-//     pub user_id: i32,
-//     pub event_id: i32,
+//     pub id: i64,
+//     pub user_id: i64,
+//     pub event_id: i64,
 //     pub created_at: SystemTime,
 //     pub updated_at: SystemTime,
 // }
 // #[derive(Debug, Insertable)]
 // #[diesel(table_name = user_events)]
 // pub struct NewUserEvent {
-//     pub user_id: i32,
-//     pub event_id: i32,
+//     pub user_id: i64,
+//     pub event_id: i64,
 // }
