+++ backend/src/models/post_models.rs
@@ -5,7 +5,7 @@ use diesel::*;
 
 use crate::{schema::{posts, user_posts, group_posts}, db_connection::PgPooledConnection};
 
-pub fn get_post(post_id: i32, conn: &mut PgPooledConnection,) -> Result<Post, Status> {
+pub fn get_post(post_id: i64, conn: &mut PgPooledConnection,) -> Result<Post, Status> {
     posts::table
         .select(posts::all_columns)
         .filter(posts::id.eq(post_id))
@@ -13,7 +13,7 @@ pub fn get_post(post_id: i32, conn: &mut PgPooledConnection,) -> Result<Post, St
         .map_err(|_| Status::new(Code::NotFound, "post_not_found"))
 }
 
-pub fn get_group_post(group_id: i32, post_id: i32, conn: &mut PgPooledConnection,) -> Result<GroupPost, Status> {
+pub fn get_group_post(group_id: i64, post_id: i64, conn: &mut PgPooledConnection,) -> Result<GroupPost, Status> {
     group_posts::table
         .select(group_posts::all_columns)
         .filter(group_posts::group_id.eq(group_id))
@@ -24,9 +24,9 @@ pub fn get_group_post(group_id: i32, post_id: i32, conn: &mut PgPooledConnection
 
 #[derive(Debug, Queryable, Identifiable, AsChangeset)]
 pub struct Post {
-    pub id: i32,
-    pub user_id: Option<i32>,
-    pub parent_post_id: Option<i32>,
+    pub id: i64,
+    pub user_id: Option<i64>,
+    pub parent_post_id: Option<i64>,
     pub title: Option<String>,
     pub link: Option<String>,
     pub content: Option<String>,
@@ -45,9 +45,9 @@ pub struct Post {
 #[derive(Debug, Queryable, Identifiable, AsChangeset)]
 #[diesel(table_name = posts)]
 pub struct MinimalPost {
-    pub id: i32,
-    pub user_id: Option<i32>,
-    pub parent_post_id: Option<i32>,
+    pub id: i64,
+    pub user_id: Option<i64>,
+    pub parent_post_id: Option<i64>,
     pub title: Option<String>,
     pub link: Option<String>,
     pub content: Option<String>,
@@ -99,8 +99,8 @@ pub static MINIMAL_POST_COLUMNS: (
 #[derive(Debug, Insertable)]
 #[diesel(table_name = posts)]
 pub struct NewPost {
-    pub user_id: Option<i32>,
-    pub parent_post_id: Option<i32>,
+    pub user_id: Option<i64>,
+    pub parent_post_id: Option<i64>,
     pub title: Option<String>,
     pub link: Option<String>,
     pub content: Option<String>,
@@ -112,10 +112,10 @@ pub struct NewPost {
 
 #[derive(Debug, Queryable, Identifiable, AsChangeset)]
 pub struct GroupPost {
-    pub id: i32,
-    pub group_id: i32,
-    pub post_id: i32,
-    pub user_id: i32,
+    pub id: i64,
+    pub group_id: i64,
+    pub post_id: i64,
+    pub user_id: i64,
     pub group_moderation: String,
     pub created_at: SystemTime,
     pub updated_at: Option<SystemTime>,
@@ -123,23 +123,23 @@ pub struct GroupPost {
 #[derive(Debug, Insertable)]
 #[diesel(table_name = group_posts)]
 pub struct NewGroupPost {
-    pub group_id: i32,
-    pub post_id: i32,
-    pub user_id: i32,
+    pub group_id: i64,
+    pub post_id: i64,
+    pub user_id: i64,
     pub group_moderation: String,
 }
 
 #[derive(Debug, Queryable, Identifiable, AsChangeset)]
 pub struct UserPost {
-    pub id: i32,
-    pub user_id: i32,
-    pub post_id: i32,
+    pub id: i64,
+    pub user_id: i64,
+    pub post_id: i64,
     pub created_at: SystemTime,
     pub updated_at: Option<SystemTime>,
 }
 #[derive(Debug, Insertable)]
 #[diesel(table_name = user_posts)]
 pub struct NewUserPost {
-    pub user_id: i32,
-    pub post_id: i32,
+    pub user_id: i64,
+    pub post_id: i64,
 }
