+++ backend/src/rpcs/get_events.rs
@@ -320,7 +320,7 @@ fn _get_my_group_posts(user: &models::User, conn: &mut PgPooledConnection) -> Ve
 }
 
 fn _get_group_posts(
-    group_id: i32,
+    group_id: i64,
     user: &Option<models::User>,
     moderations: Vec<Moderation>,
     conn: &mut PgPooledConnection,
@@ -399,7 +399,7 @@ fn _get_group_posts(
 }
 
 fn _load_group_posts(
-    group_id: i32,
+    group_id: i64,
     moderations: Vec<Moderation>,
     conn: &mut PgPooledConnection,
 ) -> Vec<Post> {
@@ -425,7 +425,7 @@ fn _load_group_posts(
 }
 
 fn _get_user_posts(
-    user_id: i32,
+    user_id: i64,
     current_user: &Option<models::User>,
     conn: &mut PgPooledConnection,
 ) -> Vec<Post> {
